chore: add yashdev9274 (#1644)
Hey there! 
It's Yash from this side. I'm new to open source and very eager to contribute my skillset in real life projects which are available on GitHub, open source fascinates me very much and I would love to be a part of your developer community so that's why I added my name along  with my GitHub profile in the readme.